+++ msr/msr_batch.py
@@ -9,7 +9,7 @@ from .msr_bounds import msr_bounds
 
 def msr_batch_from_directory(
     path: str, quiet: bool = False
-) -> list[tuple[int, int, str]]:
+) -> list[tuple[int, int, int]]:
     """
     Computes the MSR bounds for graphs in a directory with multiprocessing.
     """
@@ -22,7 +22,7 @@ def msr_batch_from_directory(
 
 def _msr_batch_from_directory_loud(
     filenames: list[str],
-) -> list[tuple[int, int, str]]:
+) -> list[tuple[int, int, int]]:
     """Uses tqdm to show progress."""
     with multiprocessing.Pool() as pool:
         return list(
@@ -35,7 +35,7 @@ def _msr_batch_from_directory_loud(
 
 def _msr_batch_from_directory_quiet(
     filenames: list[str],
-) -> list[tuple[int, int, str]]:
+) -> list[tuple[int, int, int]]:
     """Does not use tqdm to show progress."""
     with multiprocessing.Pool() as pool:
         return list(
@@ -43,7 +43,7 @@ def _msr_batch_from_directory_quiet(
         )
 
 
-def _msr_bounds_with_id_from_file(filename: str) -> tuple[int, int, str]:
+def _msr_bounds_with_id_from_file(filename: str) -> tuple[int, int, int]:
     """Helper function for msr_batch_from_directory."""
     G = load_graph(filename)
     return _msr_bounds_with_id(G)
@@ -51,7 +51,7 @@ def _msr_bounds_with_id_from_file(filename: str) -> tuple[int, int, str]:
 
 def msr_batch(
     graphs: list[graph], quiet: bool = False
-) -> list[tuple[int, int, str]]:
+) -> list[tuple[int, int, int]]:
     """Computes the MSR bounds for a batch of graphs with multiprocessing."""
     if quiet:
         return _msr_batch_quiet(graphs)
@@ -59,7 +59,7 @@ def msr_batch(
         return _msr_batch_loud(graphs)
 
 
-def _msr_batch_loud(graphs: list[graph]) -> list[tuple[int, int, str]]:
+def _msr_batch_loud(graphs: list[graph]) -> list[tuple[int, int, int]]:
     """Uses tqdm to show progress."""
     N = len(graphs)
     with multiprocessing.Pool() as pool:
@@ -68,13 +68,13 @@ def _msr_batch_loud(graphs: list[graph]) -> list[tuple[int, int, str]]:
         )
 
 
-def _msr_batch_quiet(graphs: list[graph]) -> list[tuple[int, int, str]]:
+def _msr_batch_quiet(graphs: list[graph]) -> list[tuple[int, int, int]]:
     """Does not use tqdm to show progress."""
     with multiprocessing.Pool() as pool:
         return list(pool.imap_unordered(_msr_bounds_with_id, graphs))
 
 
-def _msr_bounds_with_id(G: graph) -> tuple[int, int, str]:
+def _msr_bounds_with_id(G: graph) -> tuple[int, int, int]:
     """Helper function for msr_batch."""
     d_lo, d_hi = msr_bounds(G)
-    return d_lo, d_hi, G.name
+    return d_lo, d_hi, hash(G)